MOve code out of tests.
